Change to Esperanto for compiling Ember's packages.
There are still a number of issues with vendor packages, which is why
they are not updated already.

Vendor package issue: using `strict` mode with Esperanto causes `this`
in the root of a module to be rewritten to `undefined`. This causes
things like backburner to break.  I still need to figure out exactly how
to work around that...