Introduce license plugin
Currently the build does not attach the license from the source repo in
the built output. While a lot of JavaScript aficionados do not care
about licensing, companies do. This provides a license plugin that can
be used downstream to pluck the LICENSE file and place it into the dist.