Remove IIFE negation
V8 heuristics around IIFE's is pretty narrow and only eagerly compiles
and parses IIFEs if they begin with a paren. When iife negation is
turned on in uglify, it removes the parens and replaces them with the !
IIFE syntax which causes V8 to lazily parse functions that are meant to
be eager. Adding this flag opts out of the ! IIFE transform. @krisselden
is also opening an issue on the uglify project to change the default.

Currently you will notice that on V8 prod builds will parse/eval faster
than min builds... which is bonkers.

Reference:
https://github.com/v8/v8/blob/master/src/parsing/parser.cc#L4272-L4307
(cherry picked from commit 425716d1d01b953c07c7b87164c47b88f714c4a8)