Invert `stripRuntimeChecks` option.
`options.includeDevHelpers` is specified when we are doing a
`ember.debug.js` build.

Before this change, when `options.includeDevHelpers` was true we would
strip the `class-call-check` helper from invocations but include the
helper itself in the build ( as part of the `external-helpers-dev.js`),
and when `optioins.includeDevHelpers` was `false` we would leave the
`class-call-check` invocations in the build but **not** include the
helper itself in the build (as part of the `external-helpers-prod.js`
file).

As you can see the relationship between `includeDevHelpers` and
`stripRuntimeChecks` must be an inverse relationship, or the production
build will error.